GMSTR-3: minor, modified data types of OrderDetail entity

+++ GameStore.DAL/Entities/OrderDetail.cs
@@ -10,7 +10,7 @@ namespace GameStore.DAL.Entities
 
         public short Quantity { get; set; }
 
-        public double Discount { get; set; }
+        public float Discount { get; set; }
 
         public int OrderId { get; set; }
